Network filter for duplicate publish and confirm_ack messages
Co-authored-by: Colin Lemahieu <clemahieu@nano.org>

Adds a probabilistic filter using 128-bit siphash keys and an array. When receiving publish and confirm_ack messages, the filters (one for each message type) is checked for duplicity of the packet. The probability of a false duplicate is marginal, but not zero, and decreases with the size of the filter. The probability of a false non-duplicate is the infinitesimal probability of a 128-bit siphash collision.

Items are not normally erased. Instead, if a new item is different from the one at the insertion index (digest % capacity), the old item is replaced.

Clearing unchecked blocks resets the publish filter (assigning zero to every item).

Reaching max capacity of the inactive votes cache resets the confirm ack filter, and clears the cache. For this reason, votes inserted from this cache are now erased.