Network filter for publish and confirm_ack messages.
Adds a probabilistic filter based on direct mapped caches, keyed by 128-bit SipHash.

When receiving publish and confirm_ack messages, the filter (one for each message type) is checked for duplicity of the packet. The probability of a false duplicate is marginal, but not zero, and decreases with the size of the filter. The probability of a false non-duplicate is the infinitesimal probability of a 128-bit siphash collision.

Items are not normally erased. Instead, if a new item is different from the one at the insertion index (digest % capacity), the old item is replaced.

There are some rare instances where the filters are cleared (zero assigned to every element):
- Clearing unchecked blocks resets the publish filter.
- Reaching max capacity of the inactive votes cache resets the confirm ack filter, and clears the cache. For this reason, votes inserted from this cache are now erased.

Uses 1MB of memory for every 64k elements.